Function parsic::combinators::fix

pub fn fix<'f, A: 'f, F, S: 'f>(fix: F) -> Parser<'f, A, S> where
    F: Fn(Parser<'f, A, S>) -> Parser<'f, A, S> + 'f, 

Combinator: fix

In Rust, closures are anonymous functions, so there is no name for us to call when we want to make it recursive. Therefore, a Y-Combinator, or fixed-point combinator fix is introduced to address this issue, making it possible to write parsers that support recursive syntax using closures.

Property

fix f = f (fix f)

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{char, CharStream};

// expr := '1' expr | '0'
let parser = fix(|parser| char('1').right(parser).or(char('0')));

let mut st = CharStream::new("1110");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('0'), res);
assert_eq!("", st.as_str());
assert_eq!(0, logs.len());