Function parsic::combinators::map

pub fn map<'f, A: 'f, B: 'f, S: Clone>(
    p: impl Parsable<Stream = S, Result = A> + 'f, 
    f: impl Fn(A) -> B + 'f
) -> Parser<'f, B, S>

Combinator: map (function ver.)

Maps the result of a parser to another value using the given function.

Properties

Should satisfy Functor laws:

Identity: map(p, |x| x) ~ p

Composition: map(p, |x| f(g(x))) ~ map(map(p, f), g)

Check out test_functor module in the source code for naive examples of above laws.

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{char, CharStream};

let parser = map(or(char('H'), char('W')), |ch: char| ch == 'H');

let mut st = CharStream::new("Hello");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(true), res);
assert_eq!("ello", st.as_str());
assert_eq!(0, logs.len());