Function parsic::combinators::map_option

pub fn map_option<'f, A: 'f, B: 'f, S: Clone>(
    p: impl Parsable<Stream = S, Result = A> + 'f, 
    f: impl Fn(A) -> Option<B> + 'f
) -> Parser<'f, B, S>

Combinator: map_option (function ver.)

Maps the result of a parser to another value using the given function that produces an Option<T>. The only difference with map is that map_option will automatically try to unwrap the Option<T> and will fail if the result is None.

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{CharStream, satisfy};

let parser = map_option(satisfy(|_| true), |ch: char| ch.to_digit(10));

let mut st = CharStream::new("817");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(8), res);
assert_eq!("17", st.as_str());
assert_eq!(0, logs.len());