Function parsic::combinators::or

pub fn or<'f, A: 'f, S: Clone>(
    p1: impl Parsable<Stream = S, Result = A> + 'f, 
    p2: impl Parsable<Stream = S, Result = A> + 'f
) -> Parser<'f, A, S>

Combinator: or (function ver.)

Alternative combinator. Accepts two parsers as arguments, if the first parser succeeds then its result is returned, otherwise the result of the second parser is returned.

Properties

Should satisfy Alternative laws.

Instances of Parser and or forms a monoid:

Left identity: or(empty(), p) ~ p

Right identity: or(p, empty()) ~ p

Associative: or(or(px, py), pz) ~ or(px, or(py, pz))

Following properties exist when empty and or interacts with pure and compose:

Left zero: compose(empty(), x) ~ empty()

Right zero: compose(pf, empty()) ~ empty()

Left distribution: compose(or(pf, pg), px) ~ or(compose(pf, px), pg.compose(px))

Right distribution: compose(pf, or(px, py)) ~ or(compose(pf, px), pf.compose(py))

Left catch: or(pure(a), x) ~ pure(a)

Check out test_alternative module for naive examples of above laws.

Examples

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{char, CharStream};

// Comsumes a character 'A' or a character 'B'
let parser = or(char('B'), char('A'));

let mut st = CharStream::new("Ahhh");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('A'), res);
assert_eq!("hhh", st.as_str());
assert_eq!(0, logs.len());