Trait parsic::combinators::AlternativeExt

pub trait AlternativeExt<'f, A: 'f, S>: Parsable<Stream = S, Result = A> {
    fn or(
        self, 
        p: impl Parsable<Stream = S, Result = A> + 'f
    ) -> Parser<'f, A, S>
    where
        S: Clone,
        Self: Sized + 'f,
    { ... }
    fn optional(self) -> Parser<'f, Option<A>, S>
    where
        A: Clone,
        S: Clone + 'f,
        Self: Sized + 'f,
    { ... }
}

Provided Methods

fn or(self, p: impl Parsable<Stream = S, Result = A> + 'f) -> Parser<'f, A, S> where
S: Clone,
Self: Sized + 'f,

Combinator: or

Alternative combinator. Accepts two parsers as arguments, if the first parser succeeds then its result is returned, otherwise the result of the second parser is returned.

Properties

Should satisfy Alternative laws.

Instances of Parser and or forms a monoid:

Left identity: empty().or(p) ~ p

Right identity: p.or(empty()) ~ p

Associative: px.or(py).or(pz) ~ px.or(py.or(pz))

Following properties exist when empty and or interacts with pure and compose:

Left zero: empty().compose(x) ~ empty()

Right zero: pf.compose(empty()) ~ empty()

Left distribution: pf.or(pg).compose(px) ~ pf.compose(px).or(pg.compose(px))

Right distribution: pf.compose(px.or(py)) ~ pf.compose(px).or(pf.compose(py))

Left catch: pure(a).or(x) ~ pure(a)

Check out test_alternative module for naive examples of above laws.

Examples

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{char, CharStream};

// Comsumes a character 'A' or a character 'B'
let parser = char('B').or(char('A'));

let mut st = CharStream::new("Ahhh");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some('A'), res);
assert_eq!("hhh", st.as_str());
assert_eq!(0, logs.len());

fn optional(self) -> Parser<'f, Option<A>, S> where
A: Clone,
S: Clone + 'f,
Self: Sized + 'f,

Combinator: optional

Apply given parser at most one time. Denote the result of the given parser p as x, then the result of p.optional() would be Some(x).

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{char, CharStream};

// Consume character 't' at most one time
let parser = char('t').optional();

let mut st1 = CharStream::new("tttql");
let mut st2 = CharStream::new("ql");
let (res1, logs1) = parser.exec(&mut st1);
let (res2, logs2) = parser.exec(&mut st2);

assert_eq!(Some(Some('t')), res1);
assert_eq!(Some(None), res2);
assert_eq!(("ttql", "ql") ,(st1.as_str(), st2.as_str()));
assert_eq!((0, 0), (logs1.len(), logs2.len()));