Trait parsic::combinators::FunctorExt

pub trait FunctorExt<'f, A: 'f, S>: Parsable<Stream = S, Result = A> {
    fn map<B: 'f>(self, f: impl Fn(A) -> B + 'f) -> Parser<'f, B, S>
    where
        S: Clone,
        Self: Sized + 'f,
    { ... }
    fn map_option<B: 'f>(
        self, 
        f: impl Fn(A) -> Option<B> + 'f
    ) -> Parser<'f, B, S>
    where
        S: Clone,
        Self: Sized + 'f,
    { ... }
    fn map_result<B: 'f, E>(
        self, 
        f: impl Fn(A) -> Result<B, E> + 'f
    ) -> Parser<'f, B, S>
    where
        E: ToString,
        S: Clone,
        Self: Sized + 'f,
    { ... }
}

Provided Methods

fn map<B: 'f>(self, f: impl Fn(A) -> B + 'f) -> Parser<'f, B, S> where
S: Clone,
Self: Sized + 'f,

Combinator: map

Maps the result of current parser to another value.

Properties

Should satisfy Functor laws:

Identity: p.map(|x| x) ~ p

Composition: p.map(|x| f(g(x))) ~ p.map(f).map(g)

Check out test_functor module for naive examples of above laws.

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::*;

let parser = char('H').or(char('W'))
                      .map(|ch: char| ch == 'H');

let mut st = CharStream::new("Hello");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(true), res);
assert_eq!("ello", st.as_str());
assert_eq!(0, logs.len());

fn map_option<B: 'f>(self, f: impl Fn(A) -> Option<B> + 'f) -> Parser<'f, B, S> where
S: Clone,
Self: Sized + 'f,

Combinator: map_option

Maps the result of a parser to another value using the given function that produces an Option<T>. The only difference with map is that map_option will automatically try to unwrap the Option<T> and will fail if the result is None.

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{CharStream, satisfy};

let parser = satisfy(|_| true).map_option(|ch: char| ch.to_digit(10));

let mut st = CharStream::new("817");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(8), res);
assert_eq!("17", st.as_str());
assert_eq!(0, logs.len());

fn map_result<B: 'f, E>(
self,
f: impl Fn(A) -> Result<B, E> + 'f
) -> Parser<'f, B, S> where
E: ToString,
S: Clone,
Self: Sized + 'f,

Combinator: map_result

Maps the result of a parser to another value using the given function that produces an Result<T, E>. The only difference with map is that map_result will automatically try to unwrap the Result<T, E>. If an Err is yeilded, map_result will log down the error message. Therefore, it requires E from Result<T, E> to implement ToString trait.

Example

use parsic::combinators::*;
use parsic::core::Parsable;
use parsic::primitives::{CharStream, satisfy};

// A parser that consumes a natural number
let parser = satisfy(|&ch| ch.is_digit(10)).some()
             .map_result(|v| v.into_iter().collect::<String>().parse::<i64>());

let mut st = CharStream::new("12345");
let (res, logs) = parser.exec(&mut st);

assert_eq!(Some(12345), res);
assert_eq!("", st.as_str());
assert_eq!(0, logs.len());