Trait parsic::core::Parsable

pub trait Parsable {
    type Stream;
    type Result;
    fn parse(
        &self, 
        stream: &mut Self::Stream, 
        logger: &mut ParseLogger
    ) -> Option<Self::Result>;

    fn exec(
        &self, 
        stream: &mut Self::Stream
    ) -> (Option<Self::Result>, ParseLogger) { ... }

    fn into_parser<'f>(self) -> Parser<'f, Self::Result, Self::Stream>
    where
        Self: Sized + 'f,
    { ... }
}

Parsable trait

Anything that is parsable should implement Parsable trait, The return types of all the combinators and combinators in this library Implement Parsable trait, meaning you can treat them as parsers and call parse() or exec() from them to parse given input.

Associated Types

type Stream

type Result

Loading content...
Required methods

fn parse(
&self,
stream: &mut Self::Stream,
logger: &mut ParseLogger
) -> Option<Self::Result>

Parse function

Loading content...
Provided methods

fn exec(&self, stream: &mut Self::Stream) -> (Option<Self::Result>, ParseLogger)

Wrapper for parse function

fn into_parser<'f>(self) -> Parser<'f, Self::Result, Self::Stream> where
Self: Sized + 'f,

Convert into a Parser

Loading content...